refactor: inline OAIToAPIConverter into OAIToAPIV2Converter
OAIToAPIConverter was an abstract base class with a single subclass
(OAIToAPIV2Converter) after V1 removal. Merge all shared logic
(proxy setup, x-gravitee extension handling, server mapping) into
OAIToAPIV2Converter and delete the parent class.

https://gravitee.atlassian.net/browse/APIM-13393
