+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management/gravitee-apim-rest-api-management-rest/src/test/java/io/gravitee/rest/api/management/rest/resource/ApisResourceTest.java
@@ -158,7 +158,6 @@ public class ApisResourceTest extends AbstractResourceTest {
             .request()
             .post(Entity.json(apiDefinition));
         assertEquals(HttpStatusCode.OK_200, response.getStatus());
-
     }
 
     @Test
